AI-Assisted-Development/imageoptimizer.app/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
Werkzeug==3.1.3
requests==2.31.0
python-dotenv==1.0.0
